Removed code duplications from ModelForm
