+++ src/modules/Model/CreateModel/ModelFormMobile.tsx
@@ -3,64 +3,32 @@ import { FormLabel } from "@/ui/atoms/FormLabel";
 import { FormActionButtons } from "@molecules/FormActionButtons";
 import React from "react";
 import { Col, Container, Form, Row } from "react-bootstrap";
-import { useFieldArray, useForm } from "react-hook-form";
-import { ModelData, ModelFormProps } from "./CreateModel.types";
+import { ModelFormProps } from "./CreateModel.types";
 import {
   AddVariable,
   ModelName,
   RemoveVariable,
   VariableTypes,
 } from "./components";
+import { useCreateModelForm } from "./hooks/useCreateModelForm";
 
 export const ModelFormMobile: React.FC<ModelFormProps> = ({
   onFormSubmit,
   onCancel,
+  form,
 }) => {
   const {
-    control,
+    onSubmit,
+    fields,
+    append,
+    remove,
     handleSubmit,
     register,
-    reset,
-    formState: { errors },
-  } = useForm<ModelData>({
-    defaultValues: {
-      modelName: "",
-      variables: [
-        {
-          name: "",
-          type: "",
-          isNullable: false,
-          defaultValue: "",
-          sampleText: "",
-        },
-      ],
-    },
+    errors,
+  } = useCreateModelForm({
+    form,
+    onFormSubmit,
   });
-
-  const { fields, append, remove } = useFieldArray({
-    control,
-    name: "variables",
-  });
-
-  const onSubmit = async (data: ModelData) => {
-    onFormSubmit(data,reset);
-  };
-
-  const handleCancel = () => {
-    reset({
-      modelName: "",
-      variables: [
-        {
-          name: "",
-          type: "string",
-          isNullable: false,
-          defaultValue: "",
-          sampleText: "",
-        },
-      ],
-    });
-    onCancel();
-  };
 
   return (
     <Container fluid>
@@ -157,7 +125,7 @@ export const ModelFormMobile: React.FC<ModelFormProps> = ({
             primaryLabel="Submit"
             secondaryLabel="Cancel"
             onSubmit={handleSubmit(onSubmit)}
-            onCancel={handleCancel}
+            onCancel={onCancel}
             isPrimaryDisabled={false}
           />
         </Form>
