+++ src/modules/Model/CreateModel/index.ts
@@ -1,3 +1,2 @@
 export * from "./CreateModel";
 export * from "./CreateModel.types";
-export * from "./CreateModelMobile";
